feat(prj-runner): exec commands in project project-directory
This adds the `project exec` and `project info` commands. `exec` will
take the arguments given and spawn a new process using them. The process
will be invoked in the directory as determined by the prj-spec
implementation in the `prj-base-directory` library crate. `info` will
display information relating to the current project.