feat: split loading functionality into discover_and_assume
The `discover_and_assume` function adds an opinionated handling of what
to do when a value isn't found. It will append directory names to the
data, config, and cache directories as subdirectories of the project
root. If the project id can't be determined it will check the config
directory for a special file.