feat: generate environment variables for projects
The `generate-env` command assists in managing environment variables for
a project. This is useful for getting a project established, for example
if you want to start using the environment variables for a project then
you can generate a .env file that may be sourced. Alternatively this can
create a directory of files named after the environment variable names
(which is the primary method I wrote this for).